--- src/app/conversation/conversation.tsx
@@ -1,5 +1,0 @@
-import "@utils/entrypoints.js";
-
-export default function ConversationPage() {
-  return <div></div>;
-}
