+++ src/utils/chat.ts
@@ -1,7 +1,4 @@
-import { CallbackManager } from "langchain/callbacks";
-import { LLMChain } from "langchain/chains";
 import { ChatOpenAI } from "langchain/chat_models/openai";
-import { EntityMemory } from "langchain/memory";
 
 const OPENAI_API_KEY = process.env.NEXT_PUBLIC_OPENAI_API_KEY;
 
@@ -10,17 +7,3 @@ export const model = new ChatOpenAI({
   streaming: true,
   temperature: 0.7,
 });
-
-export const memory = new EntityMemory({
-  llm: model,
-  chatHistoryKey: "history",
-  entitiesKey: "entities",
-});
-
-export const chain = (promptType: any) => {
-  return new LLMChain({
-    llm: model,
-    prompt: promptType,
-    memory,
-  });
-};
